+++ src/main/java/com/minecolonies/coremod/colony/jobs/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the Job Classes for the workers
+ */
+package com.minecolonies.coremod.colony.jobs;
