+++ src/main/java/com/minecolonies/coremod/colony/workorders/WorkOrderBuild.java
@@ -1,48 +1,31 @@
 package com.minecolonies.coremod.colony.workorders;
 
+import com.minecolonies.coremod.blocks.AbstractBlockHut;
 import com.minecolonies.coremod.colony.CitizenData;
 import com.minecolonies.coremod.colony.Colony;
 import com.minecolonies.coremod.colony.Structures;
 import com.minecolonies.coremod.colony.buildings.AbstractBuilding;
 import com.minecolonies.coremod.colony.buildings.BuildingBuilder;
-import com.minecolonies.coremod.colony.jobs.JobBuilder;
-import com.minecolonies.coremod.util.BlockPosUtil;
+import com.minecolonies.coremod.util.BlockUtils;
 import com.minecolonies.coremod.util.LanguageHandler;
 import com.minecolonies.coremod.util.Log;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * Represents one building order to complete.
  * Has his own structure for the building.
  */
-public class WorkOrderBuild extends AbstractWorkOrder
+public class WorkOrderBuild extends WorkOrderBuildDecoration
 {
-    private static final String TAG_BUILDING       = "building";
     private static final String TAG_UPGRADE_LEVEL  = "upgradeLevel";
     private static final String TAG_UPGRADE_NAME   = "upgrade";
-    private static final String TAG_WORKORDER_NAME = "workOrderName";
-    private static final String TAG_IS_CLEARED     = "cleared";
-    private static final String TAG_IS_REQUESTED   = "requested";
-    private static final String TAG_IS_MIRRORED    = "mirrored";
 
-    private static final String TAG_SCHEMATIC_NAME    = "structureName";
-    private static final String TAG_SCHEMATIC_MD5     = "schematicMD5";
-    private static final String TAG_BUILDING_ROTATION = "buildingRotation";
-
-    protected boolean  isMirrored;
-    protected BlockPos buildingLocation;
-    protected int      buildingRotation;
-    protected String   structureName;
-    protected String   md5;
-    protected boolean  cleared;
     private   int      upgradeLevel;
     private   String   upgradeName;
-    protected String   workOrderName;
-
-    private boolean hasSentMessageForThisWorkOrder = false;
-    private boolean requested;
 
     /**
      * Unused constructor for reflection.
@@ -67,7 +50,6 @@ public class WorkOrderBuild extends AbstractWorkOrder
         this.buildingRotation = building.getRotation();
         this.isMirrored = building.getTileEntity() == null ? building.isMirrored() : building.getTileEntity().isMirrored();
         this.cleared = level > 1;
-        this.requested = false;
 
         //normalize the structureName
         Structures.StructureName sn = new Structures.StructureName(Structures.SCHEMATICS_PREFIX, building.getStyle(), this.getUpgradeName());
@@ -95,16 +77,6 @@ public class WorkOrderBuild extends AbstractWorkOrder
     }
 
     /**
-     * Get the name of the work order.
-     *
-     * @return the work order name
-     */
-    public String getName()
-    {
-        return workOrderName;
-    }
-
-    /**
      * Read the WorkOrder data from the NBTTagCompound.
      *
      * @param compound NBT Tag compound.
@@ -113,37 +85,8 @@ public class WorkOrderBuild extends AbstractWorkOrder
     public void readFromNBT(@NotNull final NBTTagCompound compound)
     {
         super.readFromNBT(compound);
-        buildingLocation = BlockPosUtil.readFromNBT(compound, TAG_BUILDING);
-        final Structures.StructureName sn = new Structures.StructureName(compound.getString(TAG_SCHEMATIC_NAME));
-        structureName = sn.toString();
-        if (!(this instanceof WorkOrderBuildDecoration))
-        {
-            upgradeLevel = compound.getInteger(TAG_UPGRADE_LEVEL);
-            upgradeName = compound.getString(TAG_UPGRADE_NAME);
-        }
-
-        workOrderName = compound.getString(TAG_WORKORDER_NAME);
-        cleared = compound.getBoolean(TAG_IS_CLEARED);
-        md5 = compound.getString(TAG_SCHEMATIC_MD5);
-        if (!Structures.hasMD5(structureName))
-        {
-            // If the schematic move we can use the MD5 hash to find it
-            final Structures.StructureName newSN = Structures.getStructureNameByMD5(md5);
-            if (newSN == null)
-            {
-                Log.getLogger().error("WorkOrderBuild.readFromNBT: Could not find " + structureName);
-            }
-            else
-            {
-                Log.getLogger().warn("WorkOrderBuild.readFromNBT: replace " + sn + " by " + newSN);
-                structureName = newSN.toString();
-
-            }
-        }
-
-        buildingRotation = compound.getInteger(TAG_BUILDING_ROTATION);
-        requested = compound.getBoolean(TAG_IS_REQUESTED);
-        isMirrored = compound.getBoolean(TAG_IS_MIRRORED);
+        upgradeLevel = compound.getInteger(TAG_UPGRADE_LEVEL);
+        upgradeName = compound.getString(TAG_UPGRADE_NAME);
     }
 
     /**
@@ -155,32 +98,8 @@ public class WorkOrderBuild extends AbstractWorkOrder
     public void writeToNBT(@NotNull final NBTTagCompound compound)
     {
         super.writeToNBT(compound);
-        BlockPosUtil.writeToNBT(compound, TAG_BUILDING, buildingLocation);
-        if (!(this instanceof WorkOrderBuildDecoration))
-        {
-            compound.setInteger(TAG_UPGRADE_LEVEL, upgradeLevel);
-            compound.setString(TAG_UPGRADE_NAME, upgradeName);
-        }
-        if (workOrderName != null)
-        {
-            compound.setString(TAG_WORKORDER_NAME, workOrderName);
-        }
-        compound.setBoolean(TAG_IS_CLEARED, cleared);
-        if (md5 != null)
-        {
-            compound.setString(TAG_SCHEMATIC_MD5, md5);
-        }
-        if (structureName == null)
-        {
-            Log.getLogger().error("WorkOrderBuild.writeToNBT: structureName should not be null!!!");
-        }
-        else
-        {
-            compound.setString(TAG_SCHEMATIC_NAME, structureName);
-        }
-        compound.setInteger(TAG_BUILDING_ROTATION, buildingRotation);
-        compound.setBoolean(TAG_IS_REQUESTED, requested);
-        compound.setBoolean(TAG_IS_MIRRORED, isMirrored);
+        compound.setInteger(TAG_UPGRADE_LEVEL, upgradeLevel);
+        compound.setString(TAG_UPGRADE_NAME, upgradeName);
     }
 
     /**
@@ -195,86 +114,43 @@ public class WorkOrderBuild extends AbstractWorkOrder
         return colony.getBuilding(buildingLocation) != null;
     }
 
-    /**
-     * Attempt to fulfill the Work Order.
-     * Override this with an implementation for the Work Order to find a Citizen to perform the job
-     * <p>
-     * finds the first suitable builder for this job.
-     *
-     * @param colony The colony that owns the Work Order.
-     */
-    @Override
-    public void attemptToFulfill(@NotNull final Colony colony)
-    {
-        boolean sendMessage = true;
-        boolean hasBuilder = false;
-
-        for (@NotNull final CitizenData citizen : colony.getCitizens().values())
-        {
-            final JobBuilder job = citizen.getJob(JobBuilder.class);
-
-            if (job == null || citizen.getWorkBuilding() == null)
-            {
-                continue;
-            }
-
-            hasBuilder = true;
-
-            final int builderLevel = citizen.getWorkBuilding().getBuildingLevel();
-
-            // don't send a message if we have a valid worker that is busy.
-            if (canBuildHut(builderLevel, citizen, colony))
-            {
-                sendMessage = false;
-            }
-
-            if (job.hasWorkOrder())
-            {
-                continue;
-            }
-
-            //  A Build WorkOrder may be fulfilled by a Builder as long as any ONE of the following is true:
-            //  - The Builder's Work AbstractBuilding is built
-            //  - OR the WorkOrder is for the Builder's Work AbstractBuilding
-            //  - OR the WorkOrder is for the TownHall
-            if (canBuildHut(builderLevel, citizen, colony))
-            {
-                job.setWorkOrder(this);
-                this.setClaimedBy(citizen);
-                return;
-            }
-        }
-
-        sendBuilderMessage(colony, hasBuilder, sendMessage);
-    }
-
-    @NotNull
-    @Override
-    protected WorkOrderType getType()
-    {
-        return WorkOrderType.BUILD;
-    }
-
     @Override
     protected String getValue()
     {
         return upgradeName;
     }
 
-    /**
-     * Checks if a builder may accept this workOrder.
-     *
-     * @param builderLevel the builder level.
-     * @return true if he is able to.
-     */
-    private boolean canBuildHut(final int builderLevel, @NotNull final CitizenData citizen, @NotNull final Colony colony)
+    private static boolean isLocationTownhall(@NotNull final Colony colony, final BlockPos buildingLocation)
     {
-        return builderLevel >= upgradeLevel || builderLevel == BuildingBuilder.MAX_BUILDING_LEVEL
-                 || (citizen.getWorkBuilding() != null && citizen.getWorkBuilding().getID().equals(buildingLocation))
-                 || isLocationTownhall(colony, buildingLocation);
+        return colony.hasTownHall() && colony.getTownHall() != null && colony.getTownHall().getID().equals(buildingLocation);
     }
 
-    private void sendBuilderMessage(@NotNull final Colony colony, final boolean hasBuilder, final boolean sendMessage)
+    /**
+     * Returns the level up level of the building.
+     *
+     * @return Level after upgrade.
+     */
+    public int getUpgradeLevel()
+    {
+        return upgradeLevel;
+    }
+
+    @Override
+    protected boolean canBuild(@NotNull final CitizenData citizen)
+    {
+        //  A Build WorkOrder may be fulfilled by a Builder as long as any ONE of the following is true:
+        //  - The Builder's Work AbstractBuilding is built
+        //  - OR the WorkOrder is for the Builder's Work AbstractBuilding
+        //  - OR the WorkOrder is for the TownHall
+
+        final int builderLevel = citizen.getWorkBuilding().getBuildingLevel();
+        return builderLevel >= upgradeLevel || builderLevel == BuildingBuilder.MAX_BUILDING_LEVEL
+                 || (citizen.getWorkBuilding() != null && citizen.getWorkBuilding().getID().equals(buildingLocation))
+                 || isLocationTownhall(citizen.getColony(), buildingLocation);
+    }
+
+    @Override
+    protected void sendBuilderMessage(@NotNull final Colony colony, final boolean hasBuilder, final boolean sendMessage)
     {
         if (hasSentMessageForThisWorkOrder)
         {
@@ -296,98 +172,36 @@ public class WorkOrderBuild extends AbstractWorkOrder
         }
     }
 
-    private static boolean isLocationTownhall(@NotNull final Colony colony, final BlockPos buildingLocation)
+    @Override
+    public int getRotation(final World world)
     {
-        return colony.hasTownHall() && colony.getTownHall() != null && colony.getTownHall().getID().equals(buildingLocation);
-    }
-
-    /**
-     * Returns the level up level of the building.
-     *
-     * @return Level after upgrade.
-     */
-    public int getUpgradeLevel()
-    {
-        return upgradeLevel;
-    }
-
-    /**
-     * Returns the ID of the building (aka ChunkCoordinates).
-     *
-     * @return ID of the building.
-     */
-    public BlockPos getBuildingLocation()
-    {
-        return buildingLocation;
-    }
-
-    /**
-     * Get the name the structure for this work order.
-     *
-     * @return the internal string for this structure.
-     */
-    public String getStructureName()
-    {
-        return structureName;
-    }
-
-    /**
-     * Gets how many times this structure should be rotated.
-     *
-     * @return building rotation.
-     */
-    public int getRotation()
-    {
+        if (buildingRotation == 0 && world != null)
+        {
+            final IBlockState blockState = world.getBlockState(buildingLocation);
+            if (blockState.getBlock() instanceof AbstractBlockHut)
+            {
+                return BlockUtils.getRotationFromFacing(blockState.getValue(AbstractBlockHut.FACING));
+            }
+        }
         return buildingRotation;
     }
 
-    /**
-     * Gets whether or not the building has been cleared.
-     *
-     * @return true if the building has been cleared.
-     */
-    public boolean isCleared()
+    @Override
+    public void onCompleted(final Colony colony)
     {
-        return cleared;
+        final BlockPos buildingLocation = getBuildingLocation();
+        final AbstractBuilding building = colony.getBuilding(buildingLocation);
+        colony.onBuildingUpgradeComplete(building, getUpgradeLevel());
     }
 
-    /**
-     * Set whether or not the building has been cleared.
-     *
-     * @param cleared true if the building has been cleared.
-     */
-    public void setCleared(final boolean cleared)
+    @Override
+    public void onRemoved(final Colony colony)
     {
-        this.cleared = cleared;
-    }
-
-    /**
-     * Gets whether or not the building materials have been requested already.
-     *
-     * @return true if the materials has been requested.
-     */
-    public boolean isRequested()
-    {
-        return requested;
-    }
-
-    /**
-     * Set whether or not the building materials have been requested already.
-     *
-     * @param requested true if so.
-     */
-    public void setRequested(final boolean requested)
-    {
-        this.requested = requested;
-    }
-
-    /**
-     * Check if the workOrder should be built isMirrored.
-     *
-     * @return true if so.
-     */
-    public boolean isMirrored()
-    {
-        return isMirrored;
+        super.onRemoved(colony);
+        final AbstractBuilding building = colony.getBuilding(getBuildingLocation());
+        if (building != null)
+        {
+            building.markDirty();
+        }
     }
 }
