+++ src/main/java/com/minecolonies/coremod/colony/workorders/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the classes relative to work orders
+ */
+package com.minecolonies.coremod.colony.workorders;
